feat: (re)structure lobby module
