+++ src/game/base/animations.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 use bevy::prelude::*;
 
 use super::resources::Animations;
@@ -8,14 +10,14 @@ pub fn animate_walking(
     animations: Res<Animations>,
 ) {
     for (entity, mut player) in players.iter_mut() {
-        /*let mut transitions = AnimationTransitions::new();
+        let mut transitions = AnimationTransitions::new();
         transitions.play(&mut player, animations.animations[0], Duration::ZERO).repeat();
         transitions.play(&mut player, animations.animations[1], Duration::ZERO).repeat();
 
         commands.entity(entity)
             .insert(animations.graph.clone())
-            .insert(transitions);*/
-        commands.entity(entity).insert(animations.graph.clone());
-        player.play(animations.animations[0]).repeat();
+            .insert(transitions);
+        //commands.entity(entity).insert(animations.graph.clone());
+        //player.play(animations.animations[0]).repeat();
     }
 }
