+++ src/game/online/mod.rs
@@ -1,10 +1,10 @@
 use bevy::prelude::*;
 use client::load_world;
-use events::{DespawnPlayer, MovePlayer, PlayerAttack, PlayerJump, ReceivedWorld, RotatePlayer, ShareAttack, ShareJump, ShareMovement, ShareRotation, ShareWorld, SpawnPlayer};
+use events::*;
 use host::share_world;
 use receive_events::{move_other_players, other_players_jump, rotate_other_players};
 use resource::{ShareMovementTimer, ShareRotationTimer};
-use share_events::{share_attack, share_jump, share_movement, share_rotation};
+use share_events::{advance_timers, share_attack, share_jump, share_movement, share_rotation};
 use ysync::TcpFromClient;
 
 use crate::{ui::{chat::ChatState, lobby::LobbySocket}, AppState};
@@ -27,13 +27,26 @@ impl Plugin for GameOnlinePlugin {
             .register_type::<Health>()
             .register_type::<Npc>()
             .register_type::<GameAge>()
-            .init_state::<OnlineGame>()
+            .init_state::<OnlineState>()
+            .add_event::<ShareWorld>()
+            .add_event::<ReceivedWorld>()
+            .add_event::<SpawnPlayer>()
+            .add_event::<DespawnPlayer>()
+            .add_event::<PlayerAttack>()
+            .add_event::<ShareMovement>()
+            .add_event::<ShareRotation>()
+            .add_event::<ShareJump>()
+            .add_event::<ShareAttack>()
+            .add_event::<MovePlayer>()
+            .add_event::<RotatePlayer>()
+            .add_event::<PlayerJump>()
             .insert_resource(ShareMovementTimer(Timer::from_seconds(0.05, TimerMode::Once)))
             .insert_resource(ShareRotationTimer(Timer::from_seconds(0.1, TimerMode::Once)))
             .add_systems(OnEnter(AppState::InGame), (
                 spawn_scene,
-            ).run_if(in_state(OnlineGame::Client)))
+            ).run_if(in_state(OnlineState::Client)))
             .add_systems(Update, (
+                advance_timers,
                 share_movement.run_if(on_event::<ShareMovement>()),
                 share_rotation.run_if(on_event::<ShareRotation>()),
                 share_jump.run_if(on_event::<ShareJump>()),
@@ -44,16 +57,16 @@ impl Plugin for GameOnlinePlugin {
                 move_other_players.run_if(on_event::<MovePlayer>()),
                 rotate_other_players.run_if(on_event::<RotatePlayer>()),
                 other_players_jump.run_if(on_event::<PlayerJump>()),
-            ).run_if(not(in_state(OnlineGame::None))))
-            .add_systems(Update, return_to_lobby.run_if(not(in_state(ChatState::Open))).run_if(not(in_state(OnlineGame::None))))
-            .add_systems(Update, load_world.run_if(on_event::<ReceivedWorld>()).run_if(in_state(OnlineGame::Client)))
-            .add_systems(Update, share_world.run_if(on_event::<ShareWorld>()).run_if(in_state(OnlineGame::Host)))
+            ).run_if(not(in_state(OnlineState::None))))
+            .add_systems(Update, return_to_lobby.run_if(not(in_state(ChatState::Open))).run_if(not(in_state(OnlineState::None))))
+            .add_systems(Update, load_world.run_if(on_event::<ReceivedWorld>()).run_if(in_state(OnlineState::Client)))
+            .add_systems(Update, share_world.run_if(on_event::<ShareWorld>()).run_if(in_state(OnlineState::Host)))
             .add_systems(OnExit(AppState::InGame), set_online_state_none);
     }
 }
 
 #[derive(States, Default, Debug, Hash, Eq, PartialEq, Clone)]
-pub enum OnlineGame {
+pub enum OnlineState {
     #[default]
     None,
     Host,
@@ -74,21 +87,21 @@ pub fn spawn_scene(
 
 pub fn return_to_lobby(
     mut remote: ResMut<LobbySocket>,
-    online_state: Res<State<OnlineGame>>,
+    online_state: Res<State<OnlineState>>,
     mut next_state: ResMut<NextState<AppState>>,
     input: Res<ButtonInput<KeyCode>>,
 ) {
     if input.just_pressed(KeyCode::Escape) {
-        if *online_state.get() == OnlineGame::Host {
+        if *online_state.get() == OnlineState::Host {
             let _ = remote.socket.tcp_send.send(TcpFromClient::GameDeletion);
         } else {
             let _ = remote.socket.tcp_send.send(TcpFromClient::GameExit);
         }
-        next_state.set(AppState::MultiplayerLobby(crate::LobbyState::InLobby));
+        next_state.set(AppState::Lobby(crate::LobbyState::InLobby));
         remote.socket.game_id = None;
     }
 }
 
-pub fn set_online_state_none(mut next_state: ResMut<NextState<OnlineGame>>) {
-    next_state.set(OnlineGame::None);
+pub fn set_online_state_none(mut next_state: ResMut<NextState<OnlineState>>) {
+    next_state.set(OnlineState::None);
 }
