+++ src/ui/components.rs
@@ -1,0 +1,8 @@
+use bevy::prelude::*;
+
+#[derive(Component)]
+pub struct NameInput;
+#[derive(Component)]
+pub struct JoinButton;
+#[derive(Component)]
+pub struct ReturnButton;
