+++ src/ui/lobby/load_game.rs
@@ -1,26 +1,26 @@
 use bevy::prelude::*;
 use ysync::TcpFromClient;
 
-use crate::{game::OnlineGame, ui::{despawn_camera, despawn_menu, spawn_camera, MenuData, HOVERED_BUTTON, NORMAL_BUTTON, PRESSED_BUTTON}, AppState, LobbyState};
+use crate::{game::online::OnlineState, ui::{components::ReturnButton, despawn_camera, despawn_menu, spawn_camera, MenuData, HOVERED_BUTTON, NORMAL_BUTTON, PRESSED_BUTTON}, AppState, LobbyState};
 
-use super::{con_selection::ReturnButton, LobbySocket};
+use super::LobbySocket;
 
 pub struct GameLoadPlugin;
 
 impl Plugin for GameLoadPlugin {
     fn build(&self, app: &mut App) {
         app
-            .add_systems(OnEnter(AppState::MultiplayerLobby(LobbyState::AwaitingJoinPermission)), (
+            .add_systems(OnEnter(AppState::Lobby(LobbyState::LoadGame)), (
                 build_load_screen,
                 spawn_camera,
             ))
-            .add_systems(OnExit(AppState::MultiplayerLobby(LobbyState::AwaitingJoinPermission)), (
+            .add_systems(OnExit(AppState::Lobby(LobbyState::LoadGame)), (
                 despawn_menu,
                 despawn_camera,
             ))
             .add_systems(Update, (
                 cancel_interaction,
-            ).run_if(in_state(AppState::MultiplayerLobby(LobbyState::AwaitingJoinPermission))));
+            ).run_if(in_state(AppState::Lobby(LobbyState::LoadGame))));
     }
 }
 
@@ -73,17 +73,17 @@ fn build_load_screen(mut commands: Commands) {
 fn cancel_interaction(
     mut next_state: ResMut<NextState<AppState>>,
     mut return_interaction_query: Query<(&Interaction, &mut BackgroundColor), (Changed<Interaction>, With<ReturnButton>)>,
-    mut online_state: ResMut<NextState<OnlineGame>>,
+    mut online_state: ResMut<NextState<OnlineState>>,
     mut remote: ResMut<LobbySocket>,
 ) {
     for (interaction, mut color) in &mut return_interaction_query {
         match *interaction {
             Interaction::Pressed => {
                 *color = PRESSED_BUTTON.into();
-                next_state.set(AppState::MultiplayerLobby(LobbyState::InLobby));
+                next_state.set(AppState::Lobby(LobbyState::InLobby));
                 let _ = remote.socket.tcp_send.send(TcpFromClient::GameExit);
                 remote.socket.game_id = None;
-                online_state.set(OnlineGame::Client);
+                online_state.set(OnlineState::Client);
             }
             Interaction::Hovered => {
                 *color = HOVERED_BUTTON.into();
